feat(takum): implement linear takum arithmetic per arXiv:2404.18603 (#593)
* feat(takum): implement linear takum arithmetic per arXiv:2404.18603

Rewrite the takum skeleton to implement the linear takum number system
specification by Hunhold (2024). The encoding uses two's complement
storage with monotonic ordering, enabling direct signed integer
comparison.

Key changes:
- Two's complement encoding: zero=0x00, NaR=0x80, negation via ~x+1
- Linear value formula: value = (-1)^sign * (1 + f) * 2^c
- Add rbits template parameter (default 3) for regime width, replacing
  hardcoded LUTs with constexpr functions dr_to_r() and dr_to_c_bias()
- Float-to-takum and takum-to-float conversions with round-to-nearest-even
- Small bit-width support: MSB-aligned truncated C field when r > nbits-overhead
- Arithmetic (+,-,*,/) via double conversion with NaR propagation
- Signed two's complement comparison operators (monotonic property)
- Exhaustive regression tests for takum<6> and takum<8> arithmetic
- Roundtrip and monotonicity verification for conversion

Closes #592


* feat(takum): add application demos for decimal accuracy and iterative refinement

Two demonstrations of takum's value in scientific computation:

1. takum_decimal_accuracy.cpp: Scans decimal accuracy (correct digits)
   across magnitudes from 10^-30 to 10^+30 for takum, posit, cfloat,
   and lns at 16 bits. Shows takum retains ~2.4 digits at all magnitudes
   while posit drops to 0 beyond 10^+/-18.

2. takum_iterative_refinement.cpp: Mixed-precision iterative refinement
   comparing 16-bit formats on two problems:
   - Hilbert matrix (near 1.0): posit advantage
   - Wide-range matrix (10^-20 to 10^+20): takum advantage
   Shows format choice depends on problem's value distribution.


* ci: add takum to conventional commit scopes


* fix(takum): resolve CodeRabbit review findings

- Fix denorm_min() to return min() when has_denorm == denorm_absent
- Use 1ull instead of 1u in shift expressions to prevent UB for rbits >= 5
- Guard operator< sign-extension against UB when nnbits == 64
- Fix monotonicity test to iterate in signed two's complement order,
  covering the negative-to-zero wrap boundary
- Add takum<6> roundtrip and monotonicity tests to regression level 1
- Fix comment/code mismatch in wide-range matrix description
- Add missing portable includes (cstdlib, cstdint, stdexcept, algorithm)
- Return zero vector on singular pivot instead of silent continue
- Guard averageDecimalAccuracy against N <= 0


* docs(takum): add encoding tables for small takum configurations

Add table generator infrastructure and pre-generated encoding tables:

- table.hpp: GenerateTakumTable<nbits,rbits>() producing TXT and CSV
  formats showing binary encoding, sign, direction, regime,
  characteristic, scale, and value for every bit pattern
- education/tables/takums.cpp: generator program for standard configs
- docs/tables/takums.txt: text tables for takum<6,3> through takum<12,3>
  plus takum<8,2> and takum<8,4> (varying regime widths)
- docs/tables/takums.csv: same in CSV format

Follows the established pattern of cfloat, posit, lns, and fixpnt tables.


* fix(takum): add ostr error check, fix Problem 2 header comment

- Add file-open error check in education/tables/takums.cpp
- Fix stale "Scaled diagonal system" comment to match actual
  "Dense wide-range matrix" construction


* fix(takum): update assignment.cpp to 3-parameter template

Fix CI build failure: assignment.cpp used old 2-parameter
takum<nbits, bt> form which conflicts with the new
takum<nbits, rbits, bt> signature.


* fix(unum2): silence sign-compare warning in operator<<

Cast left_bound to int64_t before comparing with signed expression
to eliminate -Wsign-compare warning.


* fix(takum): resolve second round of CodeRabbit review findings

- Fix NaR semantics in operator<= and operator>=: NaR comparisons
  now correctly return false (previously returned true due to
  negation of operator>/operator<)
- Guard iterative refinement against zero-norm RHS (bnorm == 0)
- Reject unsupported CLI arguments in table generator with usage msg
- Add std::exception catch clause in table generator


---------

+++ include/sw/universal/number/unum2/unum2_impl.hpp
@@ -105,7 +105,7 @@ public:
                     bound = false;
                     written = true;
 
-                    if(left_bound == static_cast<int64_t>(i - 1)) {
+                    if(static_cast<int64_t>(left_bound) == static_cast<int64_t>(i - 1)) {
                         // If inexact
                         if(left_bound & 0x01) {  // Check ubit
                             oss << "(" << u._lattice.get_exact(u._conv_idx(left_bound - 1))
